% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_topsis.R
\name{calc_topsis}
\alias{calc_topsis}
\title{Topsis multicriteria}
\usage{
calc_topsis(eachno, hazard.sc, risk.data, risk.data.hist)
}
\arguments{
\item{risk.data.hist}{}
}
\description{
Topsis multicriteria
}
