% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-coops.R
\docType{data}
\name{coops}
\alias{coops}
\title{DATASET_TITLE}
\format{
A data frame with 13 rows and 6 variables:
\describe{
\item{\code{nom_coop}}{character COLUMN_DESCRIPTION}
\item{\code{cooperativa}}{character COLUMN_DESCRIPTION}
\item{\code{localidad}}{character COLUMN_DESCRIPTION}
\item{\code{municipio}}{character COLUMN_DESCRIPTION}
\item{\code{entidad}}{character COLUMN_DESCRIPTION}
\item{\code{CVE_LOC}}{character COLUMN_DESCRIPTION}
}
}
\usage{
coops
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
