% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-coopsspecies.R
\docType{data}
\name{coopsspecies}
\alias{coopsspecies}
\title{DATASET_TITLE}
\format{
A data frame with 79 rows and 11 variables:
\describe{
\item{\code{nom_coop}}{character COLUMN_DESCRIPTION}
\item{\code{cooperativa}}{character COLUMN_DESCRIPTION}
\item{\code{esp_categoria}}{character COLUMN_DESCRIPTION}
\item{\code{nombre_comun}}{character COLUMN_DESCRIPTION}
\item{\code{nombre_cientifico}}{character COLUMN_DESCRIPTION}
\item{\code{arte_pesca}}{character COLUMN_DESCRIPTION}
\item{\code{localidad}}{character COLUMN_DESCRIPTION}
\item{\code{municipio}}{character COLUMN_DESCRIPTION}
\item{\code{entidad}}{character COLUMN_DESCRIPTION}
\item{\code{CVE_LOC}}{character COLUMN_DESCRIPTION}
\item{\code{species}}{character COLUMN_DESCRIPTION}
}
}
\usage{
coopsspecies
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
