% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-envsenspca.R
\docType{data}
\name{envsenspca}
\alias{envsenspca}
\title{DATASET_TITLE}
\format{
A data frame with 30 rows and 4 variables:
\describe{
\item{\code{norm_pca_envsens}}{double COLUMN_DESCRIPTION}
\item{\code{nom_coop}}{character COLUMN_DESCRIPTION}
\item{\code{pc_val}}{double COLUMN_DESCRIPTION}
\item{\code{scenario}}{character COLUMN_DESCRIPTION}
}
}
\usage{
envsenspca
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
