% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-expopca.R
\docType{data}
\name{expopca}
\alias{expopca}
\title{DATASET_TITLE}
\format{
A data frame with 10 rows and 3 variables:
\describe{
\item{\code{norm_pca_expo}}{double COLUMN_DESCRIPTION}
\item{\code{nom_coop}}{character COLUMN_DESCRIPTION}
\item{\code{pc_val}}{double COLUMN_DESCRIPTION}
}
}
\usage{
expopca
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
