% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-exposurehistorical.R
\docType{data}
\name{exposurehistorical}
\alias{exposurehistorical}
\title{DATASET_TITLE}
\format{
A data frame with 2219 rows and 49 variables:
\describe{
\item{\code{CVE_MUN}}{character COLUMN_DESCRIPTION}
\item{\code{CVE_LOC}}{character COLUMN_DESCRIPTION}
\item{\code{anthropogenic_index}}{double COLUMN_DESCRIPTION}
\item{\code{Average_Chlorophyll}}{double COLUMN_DESCRIPTION}
\item{\code{Average_DissolvedIron}}{double COLUMN_DESCRIPTION}
\item{\code{Average_DissolvedMolecularOxygen}}{double COLUMN_DESCRIPTION}
\item{\code{Average_Nitrate}}{double COLUMN_DESCRIPTION}
\item{\code{Average_OceanTemperature}}{double COLUMN_DESCRIPTION}
\item{\code{Average_pH}}{double COLUMN_DESCRIPTION}
\item{\code{Average_Salinity}}{double COLUMN_DESCRIPTION}
\item{\code{Average_SeaWaterDirection}}{double COLUMN_DESCRIPTION}
\item{\code{Average_SeaWaterSpeed}}{double COLUMN_DESCRIPTION}
\item{\code{Average_Silicate}}{double COLUMN_DESCRIPTION}
\item{\code{Average_TotalPhytoplankton}}{double COLUMN_DESCRIPTION}
\item{\code{biodiversity_impact}}{double COLUMN_DESCRIPTION}
\item{\code{elevation}}{double COLUMN_DESCRIPTION}
\item{\code{long_term_precipitation}}{double COLUMN_DESCRIPTION}
\item{\code{long_term_temperature}}{double COLUMN_DESCRIPTION}
\item{\code{mangrove_change}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_Chlorophyll}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_DissolvedIron}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_DissolvedMolecularOxygen}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_Nitrate}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_OceanTemperature}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_pH}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_Salinity}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_SeaWaterDirection}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_SeaWaterSpeed}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_Silicate}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_TotalPhytoplankton}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_Chlorophyll}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_DissolvedIron}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_DissolvedMolecularOxygen}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_Nitrate}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_OceanTemperature}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_pH}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_Salinity}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_SeaWaterDirection}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_SeaWaterSpeed}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_Silicate}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_TotalPhytoplankton}}{double COLUMN_DESCRIPTION}
\item{\code{red_tides}}{double COLUMN_DESCRIPTION}
\item{\code{worldclim_historical_ann_precip_1970-2000}}{double COLUMN_DESCRIPTION}
\item{\code{worldclim_historical_annual_mean_temp_1970-2000}}{double COLUMN_DESCRIPTION}
\item{\code{worldclim_historical_max_temp_1970-2000}}{double COLUMN_DESCRIPTION}
\item{\code{worldclim_historical_mean_diurnal_range_1970-2000}}{double COLUMN_DESCRIPTION}
\item{\code{worldclim_historical_min_temp_cold_month_1970-2000}}{double COLUMN_DESCRIPTION}
\item{\code{worldclim_historical_precip_driest_month_1970-2000}}{double COLUMN_DESCRIPTION}
\item{\code{worldclim_historical_precip_wettest_month_1970-2000}}{double COLUMN_DESCRIPTION}
}
}
\usage{
exposurehistorical
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
