% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-hazpca.R
\docType{data}
\name{hazpca}
\alias{hazpca}
\title{DATASET_TITLE}
\format{
A data frame with 30 rows and 3 variables:
\describe{
\item{\code{nom_coop}}{character COLUMN_DESCRIPTION}
\item{\code{scenario}}{character COLUMN_DESCRIPTION}
\item{\code{norm_pca_haz}}{double COLUMN_DESCRIPTION}
}
}
\usage{
hazpca
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
