% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lifehistorytraits.R
\docType{data}
\name{lifehistorytraits}
\alias{lifehistorytraits}
\title{DATASET_TITLE}
\format{
A data frame with 182 rows and 7 variables:
\describe{
\item{\code{species}}{character COLUMN_DESCRIPTION}
\item{\code{traits}}{character COLUMN_DESCRIPTION}
\item{\code{value}}{character COLUMN_DESCRIPTION}
\item{\code{Type}}{character COLUMN_DESCRIPTION}
\item{\code{topsis_weight}}{integer COLUMN_DESCRIPTION}
\item{\code{category}}{character COLUMN_DESCRIPTION}
\item{\code{score}}{double COLUMN_DESCRIPTION}
}
}
\usage{
lifehistorytraits
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
