% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-localities.R
\docType{data}
\name{localities}
\alias{localities}
\title{DATASET_TITLE}
\format{
A data frame with 7 rows and 9 variables:
\describe{
\item{\code{CVE_ENT}}{character COLUMN_DESCRIPTION}
\item{\code{CVE_LOC}}{character COLUMN_DESCRIPTION}
\item{\code{CVE_MUN}}{character COLUMN_DESCRIPTION}
\item{\code{NOM_ENT}}{character COLUMN_DESCRIPTION}
\item{\code{NOM_MUN}}{character COLUMN_DESCRIPTION}
\item{\code{NOM_LOC}}{character COLUMN_DESCRIPTION}
\item{\code{POBTOT}}{double COLUMN_DESCRIPTION}
\item{\code{deci_lon}}{double COLUMN_DESCRIPTION}
\item{\code{deci_lat}}{double COLUMN_DESCRIPTION}
}
}
\usage{
localities
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
