% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_factor_sensitivity.R
\name{pca_factor_sensitivity}
\alias{pca_factor_sensitivity}
\title{PCA factor}
\usage{
pca_factor_sensitivity(eachscenario, sensitivitysp, scenario.list)
}
\description{
Normalize from 0-1
}
\details{
INPUT: 1) scaled data

OUTPUT: 1) pca
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna@gmail.com
}
