% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_factor_vul.R
\name{pca_factor_vul}
\alias{pca_factor_vul}
\title{PCA factor environmental sensitivity}
\usage{
pca_factor_vul(vul.data)
}
\description{
Normalize from 0-1
}
\details{
INPUT: 1) scaled data

OUTPUT: 1) pca
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna@gmail.com
}
