% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_calc.R
\name{risk_calc}
\alias{risk_calc}
\title{Estimate risk}
\usage{
risk_calc(thisscenario, hazpca, vulpca, expopca, coops, sc.list)
}
\description{
Estimate risk by scenario
}
\details{
INPUT: 1) scaled hazard, exposure, vulnerability

OUTPUT: 1) risk
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna@gmail.com
}
