% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-scaledenvsens.R
\docType{data}
\name{scaledenvsens}
\alias{scaledenvsens}
\title{DATASET_TITLE}
\format{
A data frame with 1482 rows and 9 variables:
\describe{
\item{\code{nom_coop}}{character COLUMN_DESCRIPTION}
\item{\code{cooperativa}}{character COLUMN_DESCRIPTION}
\item{\code{localidad}}{character COLUMN_DESCRIPTION}
\item{\code{municipio}}{character COLUMN_DESCRIPTION}
\item{\code{entidad}}{character COLUMN_DESCRIPTION}
\item{\code{CVE_LOC}}{character COLUMN_DESCRIPTION}
\item{\code{scenario}}{character COLUMN_DESCRIPTION}
\item{\code{var_name}}{character COLUMN_DESCRIPTION}
\item{\code{value}}{double COLUMN_DESCRIPTION}
}
}
\usage{
scaledenvsens
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
