% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-scaledlf.R
\docType{data}
\name{scaledlf}
\alias{scaledlf}
\title{DATASET_TITLE}
\format{
A data frame with 79 rows and 22 variables:
\describe{
\item{\code{nom_coop}}{character COLUMN_DESCRIPTION}
\item{\code{cooperativa}}{character COLUMN_DESCRIPTION}
\item{\code{entidad}}{character COLUMN_DESCRIPTION}
\item{\code{municipio}}{character COLUMN_DESCRIPTION}
\item{\code{localidad}}{character COLUMN_DESCRIPTION}
\item{\code{CVE_LOC}}{character COLUMN_DESCRIPTION}
\item{\code{species}}{character COLUMN_DESCRIPTION}
\item{\code{nombre_comun}}{character COLUMN_DESCRIPTION}
\item{\code{maxima_edad}}{double COLUMN_DESCRIPTION}
\item{\code{edad_madurez}}{double COLUMN_DESCRIPTION}
\item{\code{fecundidad_media_huevos}}{double COLUMN_DESCRIPTION}
\item{\code{especificidad_habitat}}{double COLUMN_DESCRIPTION}
\item{\code{intervalo_temperatura}}{double COLUMN_DESCRIPTION}
\item{\code{intervalo_salinidad}}{double COLUMN_DESCRIPTION}
\item{\code{intervalo_profundidad}}{double COLUMN_DESCRIPTION}
\item{\code{especificidad_presa}}{double COLUMN_DESCRIPTION}
\item{\code{periodo_repro}}{double COLUMN_DESCRIPTION}
\item{\code{ciclo_repro}}{double COLUMN_DESCRIPTION}
\item{\code{dispersion_temprana}}{double COLUMN_DESCRIPTION}
\item{\code{edad_temprana}}{double COLUMN_DESCRIPTION}
\item{\code{sensibilidad_pH}}{double COLUMN_DESCRIPTION}
\item{\code{mobilidad_adulto}}{double COLUMN_DESCRIPTION}
}
}
\usage{
scaledlf
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
