% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sensitivitysp.R
\docType{data}
\name{sensitivitysp}
\alias{sensitivitysp}
\title{DATASET_TITLE}
\format{
A data frame with 5214 rows and 10 variables:
\describe{
\item{\code{CVE_LOC}}{character COLUMN_DESCRIPTION}
\item{\code{entidad}}{character COLUMN_DESCRIPTION}
\item{\code{municipio}}{character COLUMN_DESCRIPTION}
\item{\code{localidad}}{character COLUMN_DESCRIPTION}
\item{\code{species}}{character COLUMN_DESCRIPTION}
\item{\code{scenario}}{character COLUMN_DESCRIPTION}
\item{\code{nom_coop}}{character COLUMN_DESCRIPTION}
\item{\code{nombre_comun}}{character COLUMN_DESCRIPTION}
\item{\code{var_name}}{character COLUMN_DESCRIPTION}
\item{\code{value}}{double COLUMN_DESCRIPTION}
}
}
\usage{
sensitivitysp
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
