% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-senspca.R
\docType{data}
\name{senspca}
\alias{senspca}
\title{DATASET_TITLE}
\format{
A data frame with 30 rows and 5 variables:
\describe{
\item{\code{norm_pca_envsens}}{double COLUMN_DESCRIPTION}
\item{\code{nom_coop}}{character COLUMN_DESCRIPTION}
\item{\code{scenario}}{character COLUMN_DESCRIPTION}
\item{\code{norm_pca_spsens}}{double COLUMN_DESCRIPTION}
\item{\code{norm_pca}}{double COLUMN_DESCRIPTION}
}
}
\usage{
senspca
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
