% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-vulcoops.R
\docType{data}
\name{vulcoops}
\alias{vulcoops}
\title{DATASET_TITLE}
\format{
A data frame with 13 rows and 85 variables:
\describe{
\item{\code{nom_coop}}{character COLUMN_DESCRIPTION}
\item{\code{cooperativa}}{character COLUMN_DESCRIPTION}
\item{\code{localidad}}{character COLUMN_DESCRIPTION}
\item{\code{municipio}}{character COLUMN_DESCRIPTION}
\item{\code{entidad}}{character COLUMN_DESCRIPTION}
\item{\code{CVE_LOC}}{character COLUMN_DESCRIPTION}
\item{\code{pav_dis}}{double COLUMN_DESCRIPTION}
\item{\code{deg_access}}{double COLUMN_DESCRIPTION}
\item{\code{PVPH_TIC}}{double COLUMN_DESCRIPTION}
\item{\code{PVPH_AEASP}}{double COLUMN_DESCRIPTION}
\item{\code{PVPH_C_SERV}}{double COLUMN_DESCRIPTION}
\item{\code{PVPH_INTER}}{double COLUMN_DESCRIPTION}
\item{\code{PVPH_CEL}}{double COLUMN_DESCRIPTION}
\item{\code{CGRAPROES}}{double COLUMN_DESCRIPTION}
\item{\code{CPSIN_RELIG}}{double COLUMN_DESCRIPTION}
\item{\code{CP12YM_SOLT}}{double COLUMN_DESCRIPTION}
\item{\code{CHOGJEF_F}}{double COLUMN_DESCRIPTION}
\item{\code{CP12YM_SEPA}}{double COLUMN_DESCRIPTION}
\item{\code{CPRESOE15}}{double COLUMN_DESCRIPTION}
\item{\code{PTOTHOG}}{double COLUMN_DESCRIPTION}
\item{\code{POCUPVIVPAR}}{double COLUMN_DESCRIPTION}
\item{\code{PROM_OCUP}}{double COLUMN_DESCRIPTION}
\item{\code{PRO_OCUP_C}}{double COLUMN_DESCRIPTION}
\item{\code{PVIVTOT}}{double COLUMN_DESCRIPTION}
\item{\code{PHOGJEF_F}}{double COLUMN_DESCRIPTION}
\item{\code{GRAPROES_F}}{double COLUMN_DESCRIPTION}
\item{\code{PPCON_DISC}}{double COLUMN_DESCRIPTION}
\item{\code{PP_15YMAS}}{double COLUMN_DESCRIPTION}
\item{\code{PP6A11_NOA}}{double COLUMN_DESCRIPTION}
\item{\code{PP_60YMAS}}{double COLUMN_DESCRIPTION}
\item{\code{PROM_HNV}}{double COLUMN_DESCRIPTION}
\item{\code{PP12YM_SEPA}}{double COLUMN_DESCRIPTION}
\item{\code{acc_dev}}{double COLUMN_DESCRIPTION}
\item{\code{acc_3health}}{double COLUMN_DESCRIPTION}
\item{\code{acc_2health}}{double COLUMN_DESCRIPTION}
\item{\code{IM_2020}}{double COLUMN_DESCRIPTION}
\item{\code{social_indx}}{double COLUMN_DESCRIPTION}
\item{\code{OVHAC}}{double COLUMN_DESCRIPTION}
\item{\code{OVPT}}{double COLUMN_DESCRIPTION}
\item{\code{OVSAE}}{double COLUMN_DESCRIPTION}
\item{\code{OVSDE}}{double COLUMN_DESCRIPTION}
\item{\code{OVSEE}}{double COLUMN_DESCRIPTION}
\item{\code{OVSREF}}{double COLUMN_DESCRIPTION}
\item{\code{ANALF}}{double COLUMN_DESCRIPTION}
\item{\code{SBASC}}{double COLUMN_DESCRIPTION}
\item{\code{edu_pcapita}}{double COLUMN_DESCRIPTION}
\item{\code{propM_Socios}}{double COLUMN_DESCRIPTION}
\item{\code{propM_Empleados}}{double COLUMN_DESCRIPTION}
\item{\code{propM_Aspirantes}}{double COLUMN_DESCRIPTION}
\item{\code{permisos}}{double COLUMN_DESCRIPTION}
\item{\code{no_organizaciones}}{double COLUMN_DESCRIPTION}
\item{\code{financiamiento}}{double COLUMN_DESCRIPTION}
\item{\code{dur_consejo}}{double COLUMN_DESCRIPTION}
\item{\code{gobernanza}}{double COLUMN_DESCRIPTION}
\item{\code{agua}}{double COLUMN_DESCRIPTION}
\item{\code{alcant_alam}}{double COLUMN_DESCRIPTION}
\item{\code{combustible}}{double COLUMN_DESCRIPTION}
\item{\code{vigilancia}}{double COLUMN_DESCRIPTION}
\item{\code{fosa_septica}}{double COLUMN_DESCRIPTION}
\item{\code{laboratorio}}{double COLUMN_DESCRIPTION}
\item{\code{servicios}}{double COLUMN_DESCRIPTION}
\item{\code{en_solar}}{double COLUMN_DESCRIPTION}
\item{\code{planta}}{double COLUMN_DESCRIPTION}
\item{\code{telefono}}{double COLUMN_DESCRIPTION}
\item{\code{pper_socias}}{double COLUMN_DESCRIPTION}
\item{\code{pper_emp}}{double COLUMN_DESCRIPTION}
\item{\code{pper_asp}}{double COLUMN_DESCRIPTION}
\item{\code{men_18a}}{double COLUMN_DESCRIPTION}
\item{\code{18-29a}}{double COLUMN_DESCRIPTION}
\item{\code{30_40_a}}{double COLUMN_DESCRIPTION}
\item{\code{41_50_a}}{double COLUMN_DESCRIPTION}
\item{\code{51_60_a}}{double COLUMN_DESCRIPTION}
\item{\code{61_mas_a}}{double COLUMN_DESCRIPTION}
\item{\code{sin_est}}{double COLUMN_DESCRIPTION}
\item{\code{primaria}}{double COLUMN_DESCRIPTION}
\item{\code{secundaria}}{double COLUMN_DESCRIPTION}
\item{\code{preparatoria}}{double COLUMN_DESCRIPTION}
\item{\code{licenciatura}}{double COLUMN_DESCRIPTION}
\item{\code{posgrado}}{double COLUMN_DESCRIPTION}
\item{\code{ingreso_pes}}{double COLUMN_DESCRIPTION}
\item{\code{ingreso_otras}}{double COLUMN_DESCRIPTION}
\item{\code{ss_directo}}{double COLUMN_DESCRIPTION}
\item{\code{ss_indirecto}}{double COLUMN_DESCRIPTION}
\item{\code{ss_sin}}{double COLUMN_DESCRIPTION}
\item{\code{viv_10m}}{double COLUMN_DESCRIPTION}
}
}
\usage{
vulcoops
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
